feat: 🎸 Reorganization of interfaces
Externilized all interfaces in one single file
