--- src/types/password.ts
@@ -1,7 +1,0 @@
-export interface GeneratePasswordProps {
-  upperCaseEnabled: boolean;
-  numbersEnabled: boolean;
-  symbolsEnabled: boolean;
-  passwordLength: number;
-  className?: string;
-}
